requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Collection
yfinance>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3